requirements.txt
Flask==3.0.0
Jinja2==3.1.2
matplotlib==3.8.2
seaborn==0.13.0
numpy==1.26.2
questionary==2.0.1
fpdf2==2.7.6
Werkzeug==3.0.1
gunicorn==21.2.0
